requirements.txt
django==2.1.11
django-rest-auth==0.9.3
django-allauth==0.39.1
fcm-django==0.3.4
coreapi==2.3.1
django-cors-headers==3.3.0
djangorestframework==3.9.1
django-sslserver==0.19
six==1.15.0